Move default associator to its own package.
Also change the signature of `getFilteredMetricDatas` to take the
associator as input.
